# openapi.json
{"openapi":"3.0.2","info":{"title":"ToDo API","description":"A simple ToDo API built with FastAPI","version":"1.0.0"},"paths":{"/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_login_post","parameters":[{"required":true,"schema":{"title":"Username","type":"string"},"name":"username","in":"query"},{"required":true,"schema":{"title":"Password","type":"string"},"name":"password","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/todos":{"post":{"tags":["todos"],"summary":"Create Todo","operationId":"create_todo_todos_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/todos/{todo_id}":{"get":{"tags":["todos"],"summary":"Read Todo","operationId":"read_todo_todos__todo_id__get","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["todos"],"summary":"Update Todo","operationId":"update_todo_todos__todo_id__put","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["todos"],"summary":"Delete Todo","operationId":"delete_todo_todos__todo_id__delete","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Todo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/todos/{todo_id}/permissions":{"post":{"tags":["todos"],"summary":"Set Todo Permissions","operationId":"set_todo_permissions_todos__todo_id__permissions_post","parameters":[{"required":true,"schema":{"title":"Todo Id","type":"integer"},"name":"todo_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoPermission"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"Todo":{"title":"Todo","required":["title","id","owner_id"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"},"id":{"title":"Id","type":"integer"},"owner_id":{"title":"Owner Id","type":"integer"}}},"TodoCreate":{"title":"TodoCreate","required":["title"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"description":{"title":"Description","type":"string"}}},"TodoPermission":{"title":"TodoPermission","required":["user_id"],"type":"object","properties":{"user_id":{"title":"User Id","type":"integer"},"can_read":{"title":"Can Read","type":"boolean","default":false},"can_update":{"title":"Can Update","type":"boolean","default":false}}},"User":{"title":"User","required":["username","id"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"id":{"title":"Id","type":"integer"}}},"UserCreate":{"title":"UserCreate","required":["username","password"],"type":"object","properties":{"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}},"tags":[{"name":"auth","description":"Operations with user authentication"},{"name":"todos","description":"CRUD operations with todos"}]}